requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
drf-spectacular==0.27.0
pillow==10.2.0